chore: sync working changes on dev4
