+++ .gitignore
@@ -2,6 +2,7 @@
 node_modules/
 *.pnp
 .pnp.js
+.pnpm-store/
 
 # Production builds
 dist/
