+++ package.json
@@ -45,11 +45,9 @@
     "fs-extra": "^11.1.1",
     "inquirer": "^9.2.11",
     "neonctl": "^2.9.2",
-    "node-fetch": "^3.3.2",
     "open": "^10.0.0",
     "ora": "^7.0.1",
     "semver": "^7.5.4",
-    "simple-git": "^3.20.0",
     "tar": "^7.4.3",
     "which": "^4.0.0"
   },
